backend: clean up twitter.js - leave only API integration concerns, add comments
