+++ backend/test/twitterUserObjectScraping.test.js
@@ -3,6 +3,7 @@ const { executeCount } = require("../src/db/pg")
 const { setupPgBossQueue } = require("../src/db/pgBoss")
 const {
   addTwitterUserObjectScrapingJobs,
+  getTwitterScreenName,
 } = require("../src/twitterUserObjectScraping")
 const { makeSampleOrgs } = require("./prepareDb")
 
@@ -20,3 +21,13 @@ test("addTwitterUserObjectScrapingJobs", async () => {
     500 // interval
   )
 }, 10000)
+
+test("getTwitterScreenName", () => {
+  const org = {
+    id: "climatescape",
+    fields: { Twitter: "https://twitter.com/climatescape" },
+  }
+  expect(getTwitterScreenName(org)).toBe("climatescape")
+  org.fields["Twitter Override"] = "https://twitter.com/climatescape1/"
+  expect(getTwitterScreenName(org)).toBe("climatescape1")
+})
